% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{data2zscore}
\alias{data2zscore}
\title{Determine Z-Score From X}
\usage{
data2zscore(dataX, meanX, sdX)
}
\arguments{
\item{dataX:}{x value (numeric).}

\item{meanX:}{mean of x (numeric).}

\item{sdX:}{standard deviation of x (numeric).}
}
\value{

numeric value: indicating the z-score value.
}
\description{
This function computes the z-score from data(x), mean(x), and sd(x).
}
\examples{
data2zscore(10, 24, 6)
data2zscore(dataX=10, meanX=24, sdX=6)
}
\keyword{z-score}

