% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{normcdf}
\alias{normcdf}
\title{Determine Percentage Between Two Z-Score}
\usage{
normcdf(leftZ, rightZ)
}
\arguments{
\item{leftZ:}{left z-score of x (numeric).}

\item{rightZ:}{right z-score of x (numeric).}
}
\value{

numeric value: indicating the percentage value.
}
\description{
This function computes percentage between two z-score values.
}
\examples{
normcdf(1.5, 2)
normcdf(leftZ=1, rightZ=2)
}
\keyword{percentage}
\keyword{z-score}

