% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{percentile2zscore}
\alias{percentile2zscore}
\title{Determine Z-Score From Percentile}
\usage{
percentile2zscore(p)
}
\arguments{
\item{percentile:}{percentile of x (numeric).}
}
\value{

numeric value: indicating the z-score value.
}
\description{
This function computes the z-score from a percentile value.
}
\examples{
percentile2zscore(.85)
}
\keyword{percentile}
\keyword{z-score}

