% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{pop.var}
\alias{pop.var}
\title{Variance (Population)}
\usage{
pop.var(x, na.rm = FALSE)
}
\arguments{
\item{x:}{a numeric vector.}

\item{na.rm:}{logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{

numeric value: indicating the variance of the population.
}
\description{
This function compute the variance of a population x.
}
\examples{
popData <- c(4, 6, 5, 5, 6, 4, 3, 2, 5, 5, 7, 8, 9, 1, 2)
pop.var(popData)
pop.var(popData, na.rm = TRUE)
}
\keyword{variance}

