% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats1.R
\name{sample.covar}
\alias{sample.covar}
\title{Coefficient of Variation (Sample)}
\usage{
sample.covar(dataX, meanX = NULL, sdX = NULL, percentage = FALSE)
}
\arguments{
\item{dataX:}{sample (numeric vector).}

\item{meanX:}{mean of x (numeric) - defaults to NULL.}

\item{sdX:}{standard deviation of x (numeric) - defaults to NULL.}

\item{percentage:}{convert the results to a percentage (TRUE, FALSE) - defaults to FALSE.}
}
\value{
numeric value: indicating the Coefficient of Variation of the sample.
}
\description{
This function compute the coefficient of variation of a sample x.
\itemize{
   \item compare variation of two or more datasets
   \item smaller cofvar indicates a smaller variation
   \item ratio of SD to Mean (SD / Mean)
}
}
\examples{
sampleData <- c(4, 6, 5, 5, 6, 4, 3, 2, 5, 5, 7, 8, 9, 1, 2)
sample.covar(popData)
sample.covar(population = popData)
sample.covar(meanX = 5.3, sdX = 0.25, percentage = TRUE)
}
\keyword{coefficient}
\keyword{variation}

