% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{zscore2data}
\alias{zscore2data}
\title{Determine X From Z-Score}
\usage{
zscore2data(dataZ, meanX, sdX)
}
\arguments{
\item{dataZ:}{z-score of x (numeric).}

\item{meanX:}{mean of x (numeric).}

\item{sdX:}{standard deviation of x (numeric).}
}
\value{

numeric value: indicating the x value.
}
\description{
This function computes x from z-score(x), mean(x), and sd(x).
}
\examples{
zscore2data(1, 24, 6)
zscore2data(dataZ=1, meanX=24, sdX=6)
}
\keyword{z-score}

