% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats1.R
\name{zscore2percentile}
\alias{zscore2percentile}
\title{Determine Percentile From Z-Score}
\usage{
zscore2percentile(dataZ)
}
\arguments{
\item{dataZ:}{z-score of x (numeric).}
}
\value{

numeric value: indicating the percentile value.
}
\description{
This function computes percentile(x) from z-score(x).
}
\examples{
zscore2percentile(1.5)
zscore2percentile(dataZ=1)
}
\keyword{percentile}
\keyword{z-score}

